Lift ChatContextProvider
